+++ src/main/kotlin/me/steven/indrev/config/IRConfig.kt
@@ -71,196 +71,196 @@ object IRConfig {
 }
 
 class Generators {
-    val coalGenerator: GeneratorConfig = GeneratorConfig(16.0, 1.5, 1000.0, Tier.MK1.io)
+    val coalGenerator: GeneratorConfig = GeneratorConfig(16.0, 1.5, 1000, Tier.MK1.io)
 
-    val solarGeneratorMk1: GeneratorConfig = GeneratorConfig(8.0, 1.5, 500.0, Tier.MK1.io)
+    val solarGeneratorMk1: GeneratorConfig = GeneratorConfig(8.0, 1.5, 500, Tier.MK1.io)
 
-    val solarGeneratorMk3: GeneratorConfig = GeneratorConfig(12.0, 1.5, 1500.0, Tier.MK3.io)
+    val solarGeneratorMk3: GeneratorConfig = GeneratorConfig(12.0, 1.5, 1500, Tier.MK3.io)
 
-    val heatGenerator: GeneratorConfig = GeneratorConfig(128.0, -1.0, 10000.0, Tier.MK4.io)
+    val heatGenerator: GeneratorConfig = GeneratorConfig(128.0, -1.0, 10000, Tier.MK4.io)
 
-    val gasGenerator: GeneratorConfig = GeneratorConfig(1.0, 1.5, 10000.0, Tier.MK4.io)
+    val gasGenerator: GeneratorConfig = GeneratorConfig(1.0, 1.5, 10000, Tier.MK4.io)
 
-    val biomassGenerator: GeneratorConfig = GeneratorConfig(64.0, 1.5, 10000.0, Tier.MK3.io)
+    val biomassGenerator: GeneratorConfig = GeneratorConfig(64.0, 1.5, 10000, Tier.MK3.io)
 
-    val steamTurbine: GeneratorConfig = GeneratorConfig(4.0, -1.0, 20000.0, Tier.MK4.io)
+    val steamTurbine: GeneratorConfig = GeneratorConfig(4.0, -1.0, 20000, Tier.MK4.io)
 }
 
 class GeneratorConfig(
     val ratio: Double = 16.0,
     val temperatureBoost: Double = 1.5,
-    override val maxEnergyStored: Double,
-    val maxOutput: Double
+    override val maxEnergyStored: Long,
+    val maxOutput: Long
 ) : IConfig
 
 class Machines {
-    val electricFurnaceMk1: HeatMachineConfig = HeatMachineConfig(4.0, 1.0, 2.0, 1000.0, Tier.MK1.io)
+    val electricFurnaceMk1: HeatMachineConfig = HeatMachineConfig(4, 1.0, 2.0, 1000, Tier.MK1.io)
 
-    val electricFurnaceMk2: HeatMachineConfig = HeatMachineConfig(8.0, 2.0, 2.0, 5000.0, Tier.MK2.io)
+    val electricFurnaceMk2: HeatMachineConfig = HeatMachineConfig(8, 2.0, 2.0, 5000, Tier.MK2.io)
 
-    val electricFurnaceMk3: HeatMachineConfig = HeatMachineConfig(16.0, 3.0, 2.0, 10000.0, Tier.MK3.io)
+    val electricFurnaceMk3: HeatMachineConfig = HeatMachineConfig(16, 3.0, 2.0, 10000, Tier.MK3.io)
 
-    val electricFurnaceMk4: HeatMachineConfig = HeatMachineConfig(64.0, 4.0, 2.0, 100000.0, Tier.MK4.io)
+    val electricFurnaceMk4: HeatMachineConfig = HeatMachineConfig(64, 4.0, 2.0, 100000, Tier.MK4.io)
 
-    val electricFurnaceFactory: HeatMachineConfig = HeatMachineConfig(64.0, 4.0, 2.0, 100000.0, 16384.0)
+    val electricFurnaceFactory: HeatMachineConfig = HeatMachineConfig(64, 4.0, 2.0, 100000, 16384)
 
-    val pulverizerMk1: HeatMachineConfig = HeatMachineConfig(4.0, 1.0, 2.0, 1000.0, Tier.MK1.io)
+    val pulverizerMk1: HeatMachineConfig = HeatMachineConfig(4, 1.0, 2.0, 1000, Tier.MK1.io)
 
-    val pulverizerMk2: HeatMachineConfig = HeatMachineConfig(8.0, 2.0, 2.0, 5000.0, Tier.MK2.io)
+    val pulverizerMk2: HeatMachineConfig = HeatMachineConfig(8, 2.0, 2.0, 5000, Tier.MK2.io)
 
-    val pulverizerMk3: HeatMachineConfig = HeatMachineConfig(16.0, 3.0, 2.0, 10000.0, Tier.MK3.io)
+    val pulverizerMk3: HeatMachineConfig = HeatMachineConfig(16, 3.0, 2.0, 10000, Tier.MK3.io)
 
-    val pulverizerMk4: HeatMachineConfig = HeatMachineConfig(64.0, 4.0, 2.0, 100000.0, Tier.MK4.io)
+    val pulverizerMk4: HeatMachineConfig = HeatMachineConfig(64, 4.0, 2.0, 100000, Tier.MK4.io)
 
-    val pulverizerFactory: HeatMachineConfig = HeatMachineConfig(64.0, 4.0, 2.0, 100000.0, 16384.0)
+    val pulverizerFactory: HeatMachineConfig = HeatMachineConfig(64, 4.0, 2.0, 100000, 16384)
 
-    val compressorMk1: HeatMachineConfig = HeatMachineConfig(4.0, 1.0, 2.0, 1000.0, Tier.MK1.io)
+    val compressorMk1: HeatMachineConfig = HeatMachineConfig(4, 1.0, 2.0, 1000, Tier.MK1.io)
 
-    val compressorMk2: HeatMachineConfig = HeatMachineConfig(8.0, 2.0, 2.0, 5000.0, Tier.MK2.io)
+    val compressorMk2: HeatMachineConfig = HeatMachineConfig(8, 2.0, 2.0, 5000, Tier.MK2.io)
 
-    val compressorMk3: HeatMachineConfig = HeatMachineConfig(16.0, 3.0, 2.0, 10000.0, Tier.MK3.io)
+    val compressorMk3: HeatMachineConfig = HeatMachineConfig(16, 3.0, 2.0, 10000, Tier.MK3.io)
 
-    val compressorMk4: HeatMachineConfig = HeatMachineConfig(64.0, 4.0, 2.0, 100000.0, Tier.MK4.io)
+    val compressorMk4: HeatMachineConfig = HeatMachineConfig(64, 4.0, 2.0, 100000, Tier.MK4.io)
 
-    val compressorFactory: HeatMachineConfig = HeatMachineConfig(64.0, 4.0, 2.0, 100000.0, 16384.0)
+    val compressorFactory: HeatMachineConfig = HeatMachineConfig(64, 4.0, 2.0, 100000, 16384)
 
-    val sawmillMk1: HeatMachineConfig = HeatMachineConfig(4.0, 1.0, 2.0, 1000.0, Tier.MK1.io)
+    val sawmillMk1: HeatMachineConfig = HeatMachineConfig(4, 1.0, 2.0, 1000, Tier.MK1.io)
 
-    val sawmillMk2: HeatMachineConfig = HeatMachineConfig(8.0, 2.0, 2.0, 5000.0, Tier.MK2.io)
+    val sawmillMk2: HeatMachineConfig = HeatMachineConfig(8, 2.0, 2.0, 5000, Tier.MK2.io)
 
-    val sawmillMk3: HeatMachineConfig = HeatMachineConfig(16.0, 3.0, 2.0, 10000.0, Tier.MK3.io)
+    val sawmillMk3: HeatMachineConfig = HeatMachineConfig(16, 3.0, 2.0, 10000, Tier.MK3.io)
 
-    val sawmillMk4: HeatMachineConfig = HeatMachineConfig(64.0, 4.0, 2.0, 100000.0, Tier.MK4.io)
+    val sawmillMk4: HeatMachineConfig = HeatMachineConfig(64, 4.0, 2.0, 100000, Tier.MK4.io)
 
-    val infuserMk1: HeatMachineConfig = HeatMachineConfig(4.0, 1.0, 2.0, 1000.0, Tier.MK1.io)
+    val infuserMk1: HeatMachineConfig = HeatMachineConfig(4, 1.0, 2.0, 1000, Tier.MK1.io)
 
-    val infuserMk2: HeatMachineConfig = HeatMachineConfig(8.0, 2.0, 2.0,5000.0, Tier.MK2.io)
+    val infuserMk2: HeatMachineConfig = HeatMachineConfig(8, 2.0, 2.0,5000, Tier.MK2.io)
 
-    val infuserMk3: HeatMachineConfig = HeatMachineConfig(16.0, 3.0, 2.0,10000.0, Tier.MK3.io)
+    val infuserMk3: HeatMachineConfig = HeatMachineConfig(16, 3.0, 2.0,10000, Tier.MK3.io)
 
-    val infuserMk4: HeatMachineConfig = HeatMachineConfig(64.0, 4.0, 2.0,100000.0, Tier.MK4.io)
+    val infuserMk4: HeatMachineConfig = HeatMachineConfig(64, 4.0, 2.0,100000, Tier.MK4.io)
 
-    val infuserFactory: HeatMachineConfig = HeatMachineConfig(64.0, 4.5, 2.0, 100000.0, 16384.0)
+    val infuserFactory: HeatMachineConfig = HeatMachineConfig(64, 4.5, 2.0, 100000, 16384)
 
-    val chopperMk1: MachineConfig = MachineConfig(16.0, 50.0, 1000.0, Tier.MK1.io)
+    val chopperMk1: MachineConfig = MachineConfig(16, 50.0, 1000, Tier.MK1.io)
 
-    val chopperMk2: MachineConfig = MachineConfig(32.0, 40.0, 5000.0, Tier.MK2.io)
+    val chopperMk2: MachineConfig = MachineConfig(32, 40.0, 5000, Tier.MK2.io)
 
-    val chopperMk3: MachineConfig = MachineConfig(64.0, 30.0, 25000.0, Tier.MK3.io)
+    val chopperMk3: MachineConfig = MachineConfig(64, 30.0, 25000, Tier.MK3.io)
 
-    val chopperMk4: MachineConfig = MachineConfig(128.0, 20.0, 50000.0, Tier.MK4.io)
+    val chopperMk4: MachineConfig = MachineConfig(128, 20.0, 50000, Tier.MK4.io)
 
-    val farmerMk1: MachineConfig = MachineConfig(16.0, 50.0, 1000.0, Tier.MK1.io)
+    val farmerMk1: MachineConfig = MachineConfig(16, 50.0, 1000, Tier.MK1.io)
 
-    val farmerMk2: MachineConfig = MachineConfig(32.0, 40.0, 5000.0, Tier.MK2.io)
+    val farmerMk2: MachineConfig = MachineConfig(32, 40.0, 5000, Tier.MK2.io)
 
-    val farmerMk3: MachineConfig = MachineConfig(64.0, 30.0, 25000.0, Tier.MK3.io)
+    val farmerMk3: MachineConfig = MachineConfig(64, 30.0, 25000, Tier.MK3.io)
 
-    val farmerMk4: MachineConfig = MachineConfig(128.0, 20.0, 50000.0, Tier.MK4.io)
+    val farmerMk4: MachineConfig = MachineConfig(128, 20.0, 50000, Tier.MK4.io)
 
-    val slaughterMk1: MachineConfig = MachineConfig(16.0, 50.0, 1000.0, Tier.MK1.io)
+    val slaughterMk1: MachineConfig = MachineConfig(16, 50.0, 1000, Tier.MK1.io)
 
-    val slaughterMk2: MachineConfig = MachineConfig(32.0, 40.0, 5000.0, Tier.MK2.io)
+    val slaughterMk2: MachineConfig = MachineConfig(32, 40.0, 5000, Tier.MK2.io)
 
-    val slaughterMk3: MachineConfig = MachineConfig(64.0, 30.0, 25000.0, Tier.MK3.io)
+    val slaughterMk3: MachineConfig = MachineConfig(64, 30.0, 25000, Tier.MK3.io)
 
-    val slaughterMk4: MachineConfig = MachineConfig(128.0, 20.0, 50000.0, Tier.MK4.io)
+    val slaughterMk4: MachineConfig = MachineConfig(128, 20.0, 50000, Tier.MK4.io)
 
-    val rancherMk1: MachineConfig = MachineConfig(16.0, 50.0, 1000.0, Tier.MK1.io)
+    val rancherMk1: MachineConfig = MachineConfig(16, 50.0, 1000, Tier.MK1.io)
 
-    val rancherMk2: MachineConfig = MachineConfig(32.0, 40.0, 5000.0, Tier.MK2.io)
+    val rancherMk2: MachineConfig = MachineConfig(32, 40.0, 5000, Tier.MK2.io)
 
-    val rancherMk3: MachineConfig = MachineConfig(64.0, 30.0, 25000.0, Tier.MK3.io)
+    val rancherMk3: MachineConfig = MachineConfig(64, 30.0, 25000, Tier.MK3.io)
 
-    val rancherMk4: MachineConfig = MachineConfig(128.0, 20.0, 50000.0, Tier.MK4.io)
+    val rancherMk4: MachineConfig = MachineConfig(128, 20.0, 50000, Tier.MK4.io)
 
-    val miner: MachineConfig = MachineConfig(64.0, 100.0, 50000.0, Tier.MK4.io)
+    val miner: MachineConfig = MachineConfig(64, 100.0, 50000, Tier.MK4.io)
 
-    val drill: Double = 256.0
+    val drill: Long = 256
 
-    val fishingMk2: MachineConfig = MachineConfig(8.0, 400.0, 50000.0, Tier.MK2.io)
+    val fishingMk2: MachineConfig = MachineConfig(8, 400.0, 50000, Tier.MK2.io)
 
-    val fishingMk3: MachineConfig = MachineConfig(16.0, 400.0, 50000.0, Tier.MK3.io)
+    val fishingMk3: MachineConfig = MachineConfig(16, 400.0, 50000, Tier.MK3.io)
 
-    val fishingMk4: MachineConfig = MachineConfig(64.0, 500.0, 50000.0, Tier.MK4.io)
+    val fishingMk4: MachineConfig = MachineConfig(64, 500.0, 50000, Tier.MK4.io)
 
-    val dirtOxygenator: MachineConfig = MachineConfig(16.0, 50.0, 1000.0, Tier.MK1.io)
+    val dirtOxygenator: MachineConfig = MachineConfig(16, 50.0, 1000, Tier.MK1.io)
 
-    val drain: MachineConfig = MachineConfig(4.0, 20.0, 1000.0, Tier.MK1.io)
+    val drain: MachineConfig = MachineConfig(4, 20.0, 1000, Tier.MK1.io)
 
-    val pump: MachineConfig = MachineConfig(4.0, 20.0, 50.0, Tier.MK1.io)
+    val pump: MachineConfig = MachineConfig(4, 20.0, 50, Tier.MK1.io)
 
-    val smelter: HeatMachineConfig = HeatMachineConfig(64.0, 4.0, 2.0,50000.0, Tier.MK4.io)
+    val smelter: HeatMachineConfig = HeatMachineConfig(64, 4.0, 2.0,50000, Tier.MK4.io)
 
-    val condenser: MachineConfig = MachineConfig(64.0, 4.0, 50000.0, Tier.MK4.io)
+    val condenser: MachineConfig = MachineConfig(64, 4.0, 50000, Tier.MK4.io)
 
-    val fluidInfuserMk1: MachineConfig = MachineConfig(4.0, 1.0, 1000.0, Tier.MK1.io)
+    val fluidInfuserMk1: MachineConfig = MachineConfig(4, 1.0, 1000, Tier.MK1.io)
 
-    val fluidInfuserMk2: MachineConfig = MachineConfig(8.0, 2.0, 5000.0, Tier.MK2.io)
+    val fluidInfuserMk2: MachineConfig = MachineConfig(8, 2.0, 5000, Tier.MK2.io)
 
-    val fluidInfuserMk3: MachineConfig = MachineConfig(16.0, 3.0, 10000.0, Tier.MK3.io)
+    val fluidInfuserMk3: MachineConfig = MachineConfig(16, 3.0, 10000, Tier.MK3.io)
 
-    val fluidInfuserMk4: MachineConfig = MachineConfig(64.0, 4.0, 100000.0, Tier.MK4.io)
+    val fluidInfuserMk4: MachineConfig = MachineConfig(64, 4.0, 100000, Tier.MK4.io)
 
-    val electrolyticSeparatorMk1: MachineConfig = MachineConfig(4.0, 1.0, 1000.0, Tier.MK1.io)
+    val electrolyticSeparatorMk1: MachineConfig = MachineConfig(4, 1.0, 1000, Tier.MK1.io)
 
-    val electrolyticSeparatorMk2: MachineConfig = MachineConfig(8.0, 2.0, 5000.0, Tier.MK2.io)
+    val electrolyticSeparatorMk2: MachineConfig = MachineConfig(8, 2.0, 5000, Tier.MK2.io)
 
-    val electrolyticSeparatorMk3: MachineConfig = MachineConfig(16.0, 3.0, 10000.0, Tier.MK3.io)
+    val electrolyticSeparatorMk3: MachineConfig = MachineConfig(16, 3.0, 10000, Tier.MK3.io)
 
-    val electrolyticSeparatorMk4: MachineConfig = MachineConfig(64.0, 4.0, 100000.0, Tier.MK4.io)
+    val electrolyticSeparatorMk4: MachineConfig = MachineConfig(64, 4.0, 100000, Tier.MK4.io)
 
-    val modularWorkbench: MachineConfig = MachineConfig(64.0, 1.0, 5000.0, Tier.MK4.io)
+    val modularWorkbench: MachineConfig = MachineConfig(64, 1.0, 5000, Tier.MK4.io)
 
-    val laser: MachineConfig = MachineConfig(4096.0, 1.0, 2500000.0, 16384.0)
+    val laser: MachineConfig = MachineConfig(4096, 1.0, 2500000, 16384)
 
-    val distiller: MachineConfig = MachineConfig(64.0, 4.0, 100000.0, Tier.MK4.io)
+    val distiller: MachineConfig = MachineConfig(64, 4.0, 100000, Tier.MK4.io)
 
-    val lazuliFluxContainerMk1: LFCConfig = LFCConfig(10000.0, 128.0, 128.0)
+    val lazuliFluxContainerMk1: LFCConfig = LFCConfig(10000, 128, 128)
 
-    val lazuliFluxContainerMk2: LFCConfig = LFCConfig(100000.0, 512.0, 512.0)
+    val lazuliFluxContainerMk2: LFCConfig = LFCConfig(100000, 512, 512)
 
-    val lazuliFluxContainerMk3: LFCConfig = LFCConfig(1000000.0, 4096.0, 4096.0)
+    val lazuliFluxContainerMk3: LFCConfig = LFCConfig(1000000, 4096, 4096)
 
-    val lazuliFluxContainerMk4: LFCConfig = LFCConfig(10000000.0, 16384.0, 16384.0)
+    val lazuliFluxContainerMk4: LFCConfig = LFCConfig(10000000, 16384, 16384)
 }
 
 class HeatMachineConfig(
-    override val energyCost: Double,
+    override val energyCost: Long,
     override val processSpeed: Double,
     val processTemperatureBoost: Double,
-    override val maxEnergyStored: Double,
-    override val maxInput: Double
+    override val maxEnergyStored: Long,
+    override val maxInput: Long
 ) : BasicMachineConfig
 
 open class MachineConfig(
-    override val energyCost: Double,
+    override val energyCost: Long,
     override val processSpeed: Double,
-    override val maxEnergyStored: Double,
-    override val maxInput: Double
+    override val maxEnergyStored: Long,
+    override val maxInput: Long
 ) : BasicMachineConfig
 
 interface BasicMachineConfig : IConfig {
-    val energyCost: Double
+    val energyCost: Long
     val processSpeed: Double
-    val maxInput: Double
+    val maxInput: Long
 }
 
 class LFCConfig(
-    override val maxEnergyStored: Double,
-    val maxInput: Double,
-    val maxOutput: Double
+    override val maxEnergyStored: Long,
+    val maxInput: Long,
+    val maxOutput: Long
 ): IConfig
 
 interface IConfig {
-    val maxEnergyStored: Double
+    val maxEnergyStored: Long
 }
 
 class Cables {
-    val cableMk1 = 128.0
-    val cableMk2 = 512.0
-    val cableMk3 = 4096.0
-    val cableMk4 = 16384.0
+    val cableMk1 = 128
+    val cableMk2 = 512
+    val cableMk3 = 4096
+    val cableMk4 = 16384
 
     val itemPipeMk1 = 32
     val itemPipeMk2 = 64
