+++ src/main/kotlin/me/steven/indrev/gui/screenhandlers/machines/MiningRigComputerScreenHandler.kt
@@ -1,6 +1,5 @@
 package me.steven.indrev.gui.screenhandlers.machines
 
-import dev.technici4n.fasttransferlib.api.Simulation
 import io.github.cottonmc.cotton.gui.client.ScreenDrawing
 import io.github.cottonmc.cotton.gui.widget.*
 import io.github.cottonmc.cotton.gui.widget.data.HorizontalAlignment
@@ -63,7 +62,7 @@ class MiningRigComputerScreenHandler(syncId: Int, playerInventory: PlayerInvento
             bg.setSize(70, 60)
             root.add(WText(TranslatableText("block.indrev.drill.active"), HorizontalAlignment.CENTER, 0x8080), 3.45, 1.0)
 
-            val requiredPower = properties.get<Double>(MiningRigBlockEntity.ENERGY_REQUIRED_ID).toDouble()
+            val requiredPower = properties.get<Long>(MiningRigBlockEntity.ENERGY_REQUIRED_ID).toDouble()
             when {
                 properties.get<Double>(MachineBlockEntity.ENERGY_ID) < requiredPower -> {
                     val sprite = object : WSprite(identifier("textures/gui/not_enough_power.png")) {
