+++ src/main/kotlin/me/steven/indrev/gui/tooltip/energy/EnergyTooltipData.kt
@@ -3,6 +3,6 @@ package me.steven.indrev.gui.tooltip.energy
 import me.steven.indrev.gui.tooltip.CustomTooltipData
 import net.minecraft.client.gui.tooltip.TooltipComponent
 
-data class EnergyTooltipData(val energy: Double, val maxEnergy: Double) : CustomTooltipData {
+data class EnergyTooltipData(val energy: Long, val maxEnergy: Long) : CustomTooltipData {
     override fun toComponent(): TooltipComponent = EnergyTooltipComponent(this)
 }
