+++ src/main/kotlin/me/steven/indrev/items/upgrade/Enhancer.kt
@@ -21,6 +21,6 @@ enum class Enhancer {
         fun getBuffer(provider: MachineBlockEntity<*>) = getBuffer((provider as EnhancerProvider).getEnhancers(), provider)
 
         fun getBuffer(enhancers: Map<Enhancer, Int>?, provider: EnhancerProvider)
-                = provider.getBaseValue(BUFFER) + (IRConfig.upgrades.bufferUpgradeModifier * ((enhancers?: emptyMap())[BUFFER] ?: 0))
+                = (provider.getBaseValue(BUFFER) + (IRConfig.upgrades.bufferUpgradeModifier * ((enhancers?: emptyMap())[BUFFER] ?: 0))).toLong()
     }
 }
